pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "cnintendo"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "click>=8.1",
    "pymupdf>=1.23",
    "Pillow>=10.0",
    "pydantic>=2.0",
    "anthropic>=0.40",
    "openai>=1.50",
    "sqlite-utils>=3.35",
    "python-dotenv>=1.0",
    "rich>=13.0",
]

[project.scripts]
cnintendo = "cnintendo.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
